Better movement
-Move tiles are based on the players move score and the terrain
difficulty
-Players can only move once (This will be renewed during turn changes
when turns are introduced)
-Recursion is the bane of my existance

+++ Assets/Scripts/Character.cs
@@ -13,7 +13,7 @@ public class Character : MonoBehaviour
     public float attkRange;
     public float attk;
     public float defense;
-    public float move;
+    public int move;
     public float playerNumber;
     public new string name;
     public string description;
